+++ src/components/EverestUI.tsx
@@ -2,11 +2,12 @@ import { fromJS, List } from "immutable";
 import * as moment from "moment";
 import * as React from "react";
 
+import { CmrCollection, ICmrCollection } from "../types/CmrCollection";
 import { CmrGranule, ICmrGranule } from "../types/CmrGranule";
 import { IOrderParameters, OrderParameters } from "../types/OrderParameters";
 import { OrderSubmissionParameters } from "../types/OrderSubmissionParameters";
+import { cmrCollectionRequest, cmrGranuleRequest, cmrStatusRequest } from "../utils/CMR";
 import { cmrBoxArrToSpatialSelection } from "../utils/CMR";
-import { cmrGranuleRequest, cmrStatusRequest } from "../utils/CMR";
 import { IEnvironment } from "../utils/environment";
 import { hasChanged } from "../utils/hasChanged";
 import { CmrDownBanner } from "./CmrDownBanner";
@@ -23,6 +24,7 @@ interface IEverestProps {
 }
 
 interface IEverestState {
+  cmrLoading: boolean;
   cmrResponse: List<CmrGranule>;
   cmrStatusChecked: boolean;
   cmrStatusOk: boolean;
@@ -33,12 +35,8 @@ interface IEverestState {
 export class EverestUI extends React.Component<IEverestProps, IEverestState> {
     public constructor(props: any) {
       super(props);
-      this.handleOrderParameterChange = this.handleOrderParameterChange.bind(this);
-      this.handleCmrResponse = this.handleCmrResponse.bind(this);
-      this.handleCollectionChange = this.handleCollectionChange.bind(this);
-      this.onCmrRequestFailure = this.onCmrRequestFailure.bind(this);
-      this.setSpatialSelectionToCollectionDefault = this.setSpatialSelectionToCollectionDefault.bind(this);
       this.state = {
+        cmrLoading: false,
         cmrResponse: List<CmrGranule>(),
         cmrStatusChecked: false,
         cmrStatusOk: false,
@@ -63,11 +61,18 @@ export class EverestUI extends React.Component<IEverestProps, IEverestState> {
       };
 
       cmrStatusRequest().then(onSuccess, onFailure);
+
+      if (this.props.environment.inDrupal && this.props.environment.drupalDataset) {
+        cmrCollectionRequest(this.props.environment.drupalDataset.id,
+          Number(this.props.environment.drupalDataset.version))
+          .then(this.handleCmrCollectionResponse, this.onCmrRequestFailure);
+      }
     }
 
     public shouldComponentUpdate(nextProps: IEverestProps, nextState: IEverestState) {
       const propsChanged = hasChanged(this.props, nextProps, ["environment"]);
       const stateChanged = hasChanged(this.state, nextState, [
+        "cmrLoading",
         "cmrResponse",
         "cmrStatusChecked",
         "cmrStatusOk",
@@ -79,6 +84,15 @@ export class EverestUI extends React.Component<IEverestProps, IEverestState> {
     }
 
     public render() {
+      let collectionDropdown = null;
+      if (!this.props.environment.inDrupal) {
+        collectionDropdown = (
+          <CollectionDropdown
+            onCmrRequestFailure={this.props.onCmrRequestFailure}
+            cmrStatusOk={this.props.cmrStatusOk}
+            onCollectionChange={this.handleCollectionChange} />
+        );
+      }
       return (
         <div id="everest-container">
           <div id="cmr-status">
@@ -88,12 +102,7 @@ export class EverestUI extends React.Component<IEverestProps, IEverestState> {
             />
           </div>
           <div id="collection-list">
-            <CollectionDropdown
-            onCmrRequestFailure={this.onCmrRequestFailure}
-            cmrStatusOk={this.state.cmrStatusOk}
-            environment={this.props.environment}
-            selectedCollection={this.state.orderParameters.collection}
-            onCollectionChange={this.handleCollectionChange} />
+            {collectionDropdown}
           </div>
           <div id="left-side">
             <OrderParameterInputs
@@ -106,7 +115,8 @@ export class EverestUI extends React.Component<IEverestProps, IEverestState> {
           </div>
           <div id="right-side">
             <GranuleList
-              cmrResponse={this.state.cmrResponse} />
+              cmrResponse={this.state.cmrResponse}
+              loading={this.state.cmrLoading} />
             <OrderButtons
               environment={this.props.environment}
               orderSubmissionParameters={this.state.orderSubmissionParameters}
@@ -117,26 +127,34 @@ export class EverestUI extends React.Component<IEverestProps, IEverestState> {
       );
     }
 
-    private updateGranulesFromCmr() {
-      if (!this.state.cmrStatusOk) {
+    private updateGranulesFromCmr = () => {
+      if (this.state.cmrStatusChecked && !this.state.cmrStatusOk) {
         return;
       }
-      if (this.state.orderParameters.collectionId
+      if (this.state.orderParameters.collection
+          && this.state.orderParameters.collection.id
           && this.state.orderParameters.spatialSelection
           && this.state.orderParameters.temporalFilterLowerBound
           && this.state.orderParameters.temporalFilterUpperBound) {
-        cmrGranuleRequest(
-          this.state.orderParameters.collectionId,
-          this.state.orderParameters.spatialSelection,
-          this.state.orderParameters.temporalFilterLowerBound,
-          this.state.orderParameters.temporalFilterUpperBound,
-        ).then(this.handleCmrResponse, this.onCmrRequestFailure);
+        this.handleCmrGranuleRequest();
       } else {
-        console.log("Insufficient props provided.");
+        console.warn("EverestUI.updateGranulesFromCmr: Insufficient props provided.");
       }
     }
 
-    private handleOrderParameterChange(newOrderParameters: Partial<IOrderParameters>, callback: () => void) {
+    private handleCmrGranuleRequest = () => {
+      this.setState({cmrLoading: true});
+      return cmrGranuleRequest(
+        this.state.orderParameters.collection.short_name,
+        Number(this.state.orderParameters.collection.version_id),
+        this.state.orderParameters.spatialSelection,
+        this.state.orderParameters.temporalFilterLowerBound,
+        this.state.orderParameters.temporalFilterUpperBound,
+      ).then(this.handleCmrGranuleResponse, this.onCmrRequestFailure)
+       .finally(() => this.setState({cmrLoading: false}));
+    }
+
+    private handleOrderParameterChange = (newOrderParameters: Partial<IOrderParameters>, callback: () => void) => {
       // Immutable's typing for Record is incorrect; Record#merge returns a
       // Record with the same attributes, but the type definition says it
       // returns a Map (OrderParameters is a subclass of Record)
@@ -150,7 +168,6 @@ export class EverestUI extends React.Component<IEverestProps, IEverestState> {
       if (newOrderParameters.spatialSelection) {
         orderParameters = new OrderParameters({
           collection: orderParameters.collection,
-          collectionId: orderParameters.collectionId,
           spatialSelection: newOrderParameters.spatialSelection,
           temporalFilterLowerBound: orderParameters.temporalFilterLowerBound,
           temporalFilterUpperBound: orderParameters.temporalFilterUpperBound,
@@ -166,7 +183,7 @@ export class EverestUI extends React.Component<IEverestProps, IEverestState> {
       this.setState({orderParameters}, modifiedCallback);
     }
 
-    private handleCmrResponse(response: any) {
+    private handleCmrGranuleResponse = (response: any) => {
       const cmrResponse = fromJS(response.feed.entry).map((e: ICmrGranule) => new CmrGranule(e));
 
       const granuleURs = cmrResponse.map((g: CmrGranule) => g.title);
@@ -180,23 +197,41 @@ export class EverestUI extends React.Component<IEverestProps, IEverestState> {
       this.setState({cmrResponse, orderSubmissionParameters});
     }
 
-    private handleCollectionChange(collection: any) {
+    private onCmrRequestFailure = (response: any) => {
+      this.setState({cmrStatusChecked: true, cmrStatusOk: false});
+    }
+
+    private handleCollectionChange = (collection: any) => {
       this.handleOrderParameterChange({
         collection,
-        collectionId: collection.id,
         temporalFilterLowerBound: moment(collection.time_start),
         temporalFilterUpperBound: collection.time_end ? moment(collection.time_end) : moment(),
       }, this.setSpatialSelectionToCollectionDefault);
     }
 
-    private onCmrRequestFailure(response: any) {
-      this.setState({cmrStatusChecked: true, cmrStatusOk: false});
-    }
-
-    private setSpatialSelectionToCollectionDefault() {
+    private setSpatialSelectionToCollectionDefault = () => {
       const boundingBoxes = this.state.orderParameters.collection.boxes;
       const spatialSelection = cmrBoxArrToSpatialSelection(boundingBoxes);
       // @ts-ignore
       this.handleOrderParameterChange({spatialSelection});
+    }
+
+    private handleCmrCollectionResponse = (response: any) => {
+      const cmrCollections = fromJS(response.feed.entry).map((c: ICmrCollection) => new CmrCollection(c));
+
+      if (cmrCollections.size > 1) {
+        console.warn("Multiple collections matched, using first: " + cmrCollections.toJS());
+      } else if (cmrCollections.size === 0) {
+        console.warn("No collections matched: " + this.props.environment.drupalDataset);
+      }
+
+      const collection = cmrCollections.first();
+      const spatialSelection = cmrBoxArrToSpatialSelection(collection.boxes);
+      this.handleOrderParameterChange({
+        collection,
+        spatialSelection,
+        temporalFilterLowerBound: moment(collection.time_start),
+        temporalFilterUpperBound: collection.time_end ? moment(collection.time_end) : moment(),
+      }, () => null);
     }
 }