+++ src/components/OrderButtons.tsx
@@ -6,9 +6,9 @@ import { OrderSubmissionParameters } from "../types/OrderSubmissionParameters";
 import { OrderTypes } from "../types/orderTypes";
 import { IEnvironment } from "../utils/environment";
 import { hasChanged } from "../utils/hasChanged";
+import { ConfirmationFlow } from "./ConfirmationFlow";
 import { ScriptButton } from "./ScriptButton";
 import { SubmitButton } from "./SubmitButton";
-import { ViewOrderPrompt } from "./ViewOrderPrompt";
 
 interface IOrderButtonsProps {
   cmrResponse?: List<CmrGranule>;
@@ -17,58 +17,69 @@ interface IOrderButtonsProps {
 }
 
 interface IOrderButtonsState {
-  orderSubmitResponse: any;
+  orderType?: OrderTypes;
+  showConfirmationFlow: boolean;
 }
 
 export class OrderButtons extends React.Component<IOrderButtonsProps, IOrderButtonsState> {
   public constructor(props: IOrderButtonsProps) {
     super(props);
-    this.handleSubmitOrderResponse = this.handleSubmitOrderResponse.bind(this);
     this.state = {
-      orderSubmitResponse: undefined,
+      orderType: undefined, // Don't forget to set it back after submitting order
+      showConfirmationFlow: false,
     };
   }
 
   public shouldComponentUpdate(nextProps: IOrderButtonsProps, nextState: IOrderButtonsState) {
     const propsChanged = hasChanged(this.props, nextProps, ["environment", "orderSubmissionParameters"]);
-    const stateChanged = hasChanged(this.state, nextState, ["orderSubmitResponse"]);
+    const stateChanged = hasChanged(this.state, nextState, ["showConfirmationFlow"]);
 
     return propsChanged || stateChanged;
   }
 
   public render() {
-    const orderButtonsDisabled = !this.props.orderSubmissionParameters || !this.props.environment.user;
+    const loggedOut = !this.props.environment.user;
+    const orderButtonsDisabled = !this.props.orderSubmissionParameters || loggedOut;
 
     return (
       <div id="order-buttons">
         <ScriptButton
           disabled={orderButtonsDisabled}
           environment={this.props.environment}
+          loggedOut={loggedOut}
           cmrResponse={this.props.cmrResponse} />
         <SubmitButton
           buttonText={"Get Individual Files"}
           disabled={orderButtonsDisabled}
-          environment={this.props.environment}
-          hoverText={"Once the order is processed, go to the Order page for a list of links to your files."}
-          orderSubmissionParameters={this.props.orderSubmissionParameters}
-          onSubmitOrderResponse={this.handleSubmitOrderResponse}
+          hoverText={"Once processed, your Order page will have a linked list of files."}
+          loggedOut={loggedOut}
+          onSubmitOrder={this.handleSubmitOrder}
           orderType={OrderTypes.listOfLinks} />
         <SubmitButton
           buttonText={"Order Zip File"}
           disabled={orderButtonsDisabled}
-          environment={this.props.environment}
-          hoverText={"Once the order is processed, go to the Order page for a list of links to your files."}
-          orderSubmissionParameters={this.props.orderSubmissionParameters}
-          onSubmitOrderResponse={this.handleSubmitOrderResponse}
+          hoverText={"Once processed, your Order page will link to one or more zipped files."}
+          loggedOut={loggedOut}
+          onSubmitOrder={this.handleSubmitOrder}
           orderType={OrderTypes.zipFile} />
-        <ViewOrderPrompt
+        <ConfirmationFlow
           environment={this.props.environment}
-          orderSubmitResponse={this.state.orderSubmitResponse} />
+          onRequestClose={this.closeConfirmationFlow}
+          orderSubmissionParameters={this.props.orderSubmissionParameters}
+          orderType={this.state.orderType}
+          show={this.state.showConfirmationFlow} />
       </div>
     );
   }
 
-  private handleSubmitOrderResponse(hermesResponse: any) {
-    this.setState({orderSubmitResponse: hermesResponse});
+  public closeConfirmationFlow = () => {
+    this.setState({showConfirmationFlow: false});
+  }
+
+  private handleSubmitOrder = (orderType: OrderTypes) => {
+    this.setState({
+      orderType,
+      showConfirmationFlow: true,
+    });
   }
 }